pyproject.toml
[tool.poetry]
name = "fastapi-alembic-postgres"
version = "0.1.0"
description = ""
authors = ["Quien Sabe <qs5779@mail.com>"]
readme = "README.md"
packages = [{include = "fap"}]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = {extras = ["mypy"], version = "^2.0.13"}
alembic = "^1.10.4"
fastapi = "^0.95.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
pydantic = {extras = ["dotenv", "email", "mypy"], version = "^1.10.7"}
psycopg2 = "^2.9.6"
uvicorn = "^0.22.0"
tenacity = "^8.2.2"
jinja2 = "^3.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"


[tool.poetry.group.test.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
pytest = "^7.3.1"
wemake-python-styleguide = "^0.17.0"
loguru-mypy = "^0.0.4"
types-passlib = "^1.7.7.12"
types-python-jose = "^3.3.4.7"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
